refactor: rename enams to enums
