+++ src/app/modals/enums/modal-component.enum.ts
